修改Int.dp返回值类型为Int

+++ library/src/main/java/com/github/zicheng/banner/ext/Extensions.kt
@@ -18,5 +18,5 @@ internal val Int.dp
         TypedValue.COMPLEX_UNIT_DIP,
         this.toFloat(),
         displayMetrics
-    )
+    ).toInt()
